sdk: Point/Rect/Size become records of ints (phase 3.5)
The three geometry types were org.opencv.core.* clones — mutable public
double fields, set(double[]), double[] constructors, clone(), Size(Point),
Rect.fullScreen(). Nothing in the SDK, Studio or any generated file ever
mutated one, and the vestigial half had no callers at all. The mutability
bought nothing and cost a real defect: Point and Rect declared no equals,
so p1.equals(p2) in a bot was an identity comparison — a silently wrong
answer rather than a compile error.

int rather than double because every producer is a pixel and every
consumer is an input event the native layer can only deliver at a whole
pixel; the double was carried only to be cast straight back, at fourteen
call sites. The three genuinely fractional producers — Rect's and
MatchResult's midpoints, Pixel's centre of mass, Mouse.drag's
interpolation — round where the value is built. drag now rounds per step
instead of truncating, which had biased every step of a glide toward its
start.

Immutability removes the defensive copies ColorMatch.getCenter/getTopLeft,
MatchResult.getTopLeft and TextMatch.getBounds each made to avoid handing
out a mutable field.

Done now because the contract starts at 1.1.0 and the newest tag is
v1.0.26, so this is the last release it is free in — and it is the one
break the pointer model cannot carry, since a public field becoming an
accessor turns p.x into p.x(), which is not a call and so is invisible to
SdkReferences and unrewritable by CallMigrator.

Studio needs no change: VariableWire.geometryHelper already emitted
int[] n = ints(s, N); new Point(n[0], n[1]).

+++ src/main/java/com/botmaker/sdk/api/emulator/Emulator.java
@@ -71,7 +71,7 @@ public final class Emulator implements CaptureSource {
             return;
         }
         Debug.log("[Emulator] click " + p + " on " + name());
-        device.tap((int) p.x, (int) p.y);
+        device.tap(p.x(), p.y());
     }
 
     // --- Native emulator input verbs ---
